Add root & context database instances
Using graph-crdt, a simple asynchronous Graph interface has been built
on top and exported as `database`. Key difference: `read` and `write`
methods are asynchronous, leaving room for a variety of plugins to be
called.

Changelog updated.

The Bluebird API isn't being used anywhere in mytosis. It's also
throwing all sorts of errors when prototyping with async generators, so
for now I'm dropping the library.