Split node/field reads into different pipes
All writes can be expressed as a graph. Write pipelines don't care if
that graph contains new nodes, or just changes a single node. That same
statement doesn't work for read pipelines. Reading a UID is very
different from reading a field, and oftentimes the return values are
radically different.

For this reason, the read pipeline cannot be a singular entity if it's
going to intercept field reads as well. As a solution, the pipeline has
been split into two categories: node, and field. However, this doubled
the pipeline surface area and introduced more complexity into pipeline
triggers and config generation. Significant portions of the code was
changed.