Refactor pipeline (broken state)
The pipeline now only accepts a single initial state, assumed to be an
object. This (massively) simplifies the flow through a pipeline,
improves performance, and allows pipelines to add new fields and
deprecate others without worrying about the ordering of arguments, or
how many are passed already.

Tests are currently failing. Some other points in the library depended
on it being an argument based pipeline.