Fix pipeline behavior
A substantial refactor of the pipeline happened, and the root/context
interfaces weren't updated to handle it. This caused a ton of tests
(namely the hooks) to fail. This commit updates how root and context
instances interface with the pipeline, and fixes all the broken tests to
use the newer API.