Remove dead code and reuse mocks
There were several places where mocks were used. For a while, they were
all in the database/ folder, so that's where the mocks lived. Demand
grew in other folders, so rather than linking inside the folder, they've
been moved to the src/ root.

In other news, there was some dead code in the pipeline (the `defaults`
function) which was replaced with `addDefaultProps`. Since
`addDefaultProps` was the less tested of the two, it had to go.