Pass database and settings to router
I overlooked the bit where the router will need to initiate actions with
the database, not just the other way around. Since it needs a reference,
the `config.router` property is treated as a router factory (the name
may change later to better reflect that). The datatabase instance and
config is passed as parameters.