Fix missing dep on node v4
Okay, I'm totally guessing. The build says it can't find babel runtime,
which makes sense because of the new `transform-runtime` plugin. I've
added it to the package in the hopes it fixes the build, but it seems
odd since it only failed on node v4 last time, and this wouldn't account
for that.