Prevent enums with derivations and their subtype
Prevents an edge case where developers could add a derivation to an enum
along with the base type yielding an ambiguous situation. Concrete
example: say the `time` primitive is derived from `string`. If you put
both `time` and `string` in an enum, how do you know which the value
belongs to? You don't.