Add Stream#some and dispose to observers
Observers can now invoke `dispose` from inside the observer function.
This comes at necessity because of an edge case:
- Synchronous action producer with two or more events
- Resolve stream on first successful predicate test
- The disposer isn't called because `observe()` is still executing (as
`_openStream`), so there's no return callback to invoke
- The observer is force-fed events it doesn't need until the publisher
hits an async point.

That's not ideal. As a solution, I've defined the disposer higher and
passed it as a parameter to observers, so they can unsubscribe
immediately if synchronous action creators are dangerous.