Fix buffer and composite name validation
Allows numbers in composite names, so long as they aren't the first
character.

Fixes some dumb stuff I did by getting the Uint8Array subclass instead
of checking for each of the buffer view constructors. IE 11 doesn't work
like that (which doesn't surprise me, it was a hack to begin with), and
there's a method built right into `ArrayBuffer` which handles it
automatically.