Mute Symbol.asyncIterator trash fire
Upgrading to babel 7 changed how `Symbol.asyncIterator` gets polyfilled.
Almost the entire test suite for the leveldb plugin depends on that
feature, and was failing with a really weird runtime error.

The suggested solutions are either to include babel-polyfill in the
library, or run it as part of the test framework. I might do that, or
I might procrastinate it. Up to me.