Kill off DB implementation
Apparently I _was_ using the `streams` implementation. Irrelevant. The
`@mytosis/db` project didn't work, was based on really old ideas, and
was heading in the wrong direction. The only thing I'm moderately sure
about is the type system. The rest will need to be rewritten. Deleting
the code makes it easier to manage and more explicit about the
intentions.